requirements.txt
# Core Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Vector Database
qdrant-client>=1.7.0

# LLM & RAG
langchain>=0.1.0
llama-index>=0.9.0
openai>=1.10.0
anthropic>=0.18.0
google-generativeai>=0.3.0
groq>=0.4.0
mistralai>=0.1.0
cohere>=4.0.0
httpx>=0.25.0

# Document Processing
pymupdf4llm>=0.0.5
python-docx>=1.1.0
docxtpl>=0.16.0
weasyprint>=60.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdfplumber>=0.11.0
python-magic>=0.4.27

# NLP - Arabic
camel-tools>=1.5.0

# NLP - General
sentence-transformers>=2.3.0
langdetect>=1.0.9
transformers>=4.45.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Caching
redis>=5.0.0

# Utilities
numpy>=1.26.0
pandas>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0
